requirements.txt
selenium==3.141.0
requests==2.26.0
beautifulsoup4==4.10.0
python_dateutil==2.8.2
